backend/requirements.txt
Package             Version
------------------- -------
asgiref             3.5.2
autopep8            1.6.0
Django              4.0.6
django-cors-headers 3.13.0
djangorestframework 3.13.1
djongo              1.3.6
dnspython           2.2.1
pip                 22.1.2
pycodestyle         2.8.0
pymongo             3.12.3
pytz                2022.1
setuptools          62.1.0
sqlparse            0.2.4
toml                0.10.2
tzdata              2022.1
wheel               0.37.1
